+++ kamehouse-shell/bin/kamehouse/tomcat/tomcat-restart-dev.sh
@@ -30,7 +30,7 @@ killRemainingTomcatProcess() {
       log.info "Tomcat is not running on port ${TOMCAT_PORT}"
     else
       log.info "Killing process ${COL_PURPLE}${TOMCAT_PID}"
-      cmd.exe "/c taskkill.exe /PID ${TOMCAT_PID} /F"
+      powershell.exe -c "taskkill.exe /PID ${TOMCAT_PID} /F"
       powershell.exe -c "Stop-Process -Id ${TOMCAT_PID} -Force"
     fi
   fi