+++ kamehouse-shell/bin/win/kamehouse/httpd-stop.sh
@@ -15,7 +15,7 @@ mainProcess() {
     log.info "Apache httpd is not running"
   else
     log.info "Killing process ${COL_PURPLE}${HTTPD_PID}"
-    cmd.exe "/c taskkill.exe /PID ${HTTPD_PID} /F"
+    powershell.exe -c "taskkill.exe /PID ${HTTPD_PID} /F"
     powershell.exe -c "Stop-Process -Id ${HTTPD_PID} -Force"
   fi
 }