+++ kamehouse-shell/bin/win/shutdown/shutdown-status.sh
@@ -4,7 +4,7 @@ source ${HOME}/programs/kamehouse-shell/bin/common/functions/common-functions.sh
 if [ "$?" != "0" ]; then echo "Error importing common-functions.sh" ; exit 99 ; fi
 
 mainProcess() {
-  cmd.exe "/c tasklist.exe /FI IMAGENAME eq shutdown.exe"
+  powershell.exe -c "tasklist.exe /FI IMAGENAME eq shutdown.exe"
 }
 
 main "$@"