+++ kamehouse-shell/bin/win/shutdown/suspend.sh
@@ -4,7 +4,7 @@ source ${HOME}/programs/kamehouse-shell/bin/common/functions/kamehouse/kamehouse
 if [ "$?" != "0" ]; then echo "Error importing kamehouse-functions.sh" ; exit 99 ; fi
 
 mainProcess() {
-  cmd.exe "/c rundll32.exe powrprof.dll,SetSuspendState 0,1,0"
+  powershell.exe -c "rundll32.exe powrprof.dll,SetSuspendState 0,1,0"
 }
 
 main "$@"